+++ oservi/engines/__init__.py
@@ -29,6 +29,12 @@ from oservi.engines import feed_tracker as _feed_tracker  # noqa: F401
 from oservi.engines import sequential_composer as _sequential_composer  # noqa: F401
 from oservi.engines import subagent_orchestrator as _subagent_orchestrator  # noqa: F401
 from oservi.engines import mcp_bridge as _mcp_bridge  # noqa: F401
+from oservi.engines import saga_composer as _saga_composer  # noqa: F401
+from oservi.engines import state_machine_engine as _state_machine_engine  # noqa: F401
+from oservi.engines import event_webhook_dispatcher as _event_webhook_dispatcher  # noqa: F401
+from oservi.engines import cron_scheduler_engine as _cron_scheduler_engine  # noqa: F401
+from oservi.engines import bulk_import_worker as _bulk_import_worker  # noqa: F401
+from oservi.engines import bulk_export_worker as _bulk_export_worker  # noqa: F401
 
 from oservi.engines.alerter import AlerterEngine
 from oservi.engines.researcher import ResearcherEngine
@@ -36,6 +42,12 @@ from oservi.engines.feed_tracker import FeedTrackerEngine
 from oservi.engines.sequential_composer import SequentialComposerEngine
 from oservi.engines.subagent_orchestrator import SubagentOrchestratorEngine
 from oservi.engines.mcp_bridge import McpBridgeEngine
+from oservi.engines.saga_composer import SagaComposerEngine
+from oservi.engines.state_machine_engine import StateMachineEngine
+from oservi.engines.event_webhook_dispatcher import EventWebhookDispatcherEngine
+from oservi.engines.cron_scheduler_engine import CronSchedulerEngine
+from oservi.engines.bulk_import_worker import BulkImportWorkerEngine
+from oservi.engines.bulk_export_worker import BulkExportWorkerEngine
 
 __all__ = [
     "EngineSkeleton",
@@ -49,4 +61,10 @@ __all__ = [
     "SequentialComposerEngine",
     "SubagentOrchestratorEngine",
     "McpBridgeEngine",
+    "SagaComposerEngine",
+    "StateMachineEngine",
+    "EventWebhookDispatcherEngine",
+    "CronSchedulerEngine",
+    "BulkImportWorkerEngine",
+    "BulkExportWorkerEngine",
 ]